restructured methods
