+++ scripts/performance/diffusion/pretrain_flux_12b.py
@@ -20,7 +20,7 @@ from nemo.collections.diffusion.recipes.flux_12b import pretrain_recipe
 from nemo.lightning.run.plugins import NsysPlugin, PerfEnvPlugin
 
 from ..argument_parser import parse_cli_args
-from ..utils import (
+from ..helpers import (
     args_sanity_check,
     get_user_configs,
     set_exp_logging_configs,
