+++ scripts/performance/llm/finetune_llama31_405b.py
@@ -27,16 +27,18 @@ from nemo.collections.nlp.modules.common.tokenizer_utils import get_nmt_tokenize
 from nemo.lightning.run.plugins import MemoryProfilePlugin, NsysPlugin, PerfEnvPlugin
 
 from ..argument_parser import parse_cli_args
-from ..utils import (
+from ..helpers import (
     args_sanity_check,
-    get_comm_overlap_callback_idx,
     get_user_configs,
-    hf_tokenizer,
-    import_ckpt_experiment,
-    isfile_train_pack_metadata,
     set_exp_logging_configs,
     set_primary_perf_configs,
     slurm_executor,
+)
+from ..utils import (
+    hf_tokenizer,
+    import_ckpt_experiment,
+    isfile_train_pack_metadata,
+    get_comm_overlap_callback_idx,
 )
 
 HF_MODEL_URI = "meta-llama/Llama-3.1-405B"
