+++ scripts/performance/llm/mlperf_lora_llama2_70b.py
@@ -27,7 +27,7 @@ from nemo.lightning.pytorch.optim import CosineAnnealingScheduler
 from nemo.lightning.run.plugins import MemoryProfilePlugin, NsysPlugin, PerfEnvPlugin
 
 from ..argument_parser import parse_cli_args
-from ..utils import args_sanity_check, import_ckpt_experiment, slurm_executor
+from ..helpers import args_sanity_check, import_ckpt_experiment, slurm_executor
 
 NUM_NODES = 1
 NUM_GPUS_PER_NODE = 8
