+++ scripts/performance/llm/pretrain_nemotron4_15b.py
@@ -24,15 +24,15 @@ from nemo.collections.nlp.modules.common.tokenizer_utils import get_nmt_tokenize
 from nemo.lightning.run.plugins import MemoryProfilePlugin, NsysPlugin, PerfEnvPlugin
 
 from ..argument_parser import parse_cli_args
-from ..utils import (
+from ..helpers import (
     args_sanity_check,
-    get_comm_overlap_callback_idx,
     get_user_configs,
     logging,
     set_exp_logging_configs,
     set_primary_perf_configs,
     slurm_executor,
 )
+from ..utils import get_comm_overlap_callback_idx
 
 
 def override_recipe_configs(
