+++ scripts/performance/vlm/finetune_neva_8b.py
@@ -21,7 +21,7 @@ from nemo.collections.vlm.recipes.neva_llama3_8b import finetune_recipe
 from nemo.lightning.run.plugins import NsysPlugin, PerfEnvPlugin
 
 from ..argument_parser import parse_cli_args
-from ..utils import (
+from ..helpers import (
     args_sanity_check,
     get_user_configs,
     set_exp_logging_configs,
